Compiler warning cleanup

+++ SRWebSocketTests/SRTAutobahnTests.m
@@ -139,7 +139,7 @@
         return;
     }
     [aRun start];
-    for (int i = 1; i <= aRun.test.testCaseCount; i++) {
+    for (NSUInteger i = 1; i <= aRun.test.testCaseCount; i++) {
         SEL sel = @selector(performTestWithNumber:);
         NSInvocation *invocation = [NSInvocation invocationWithMethodSignature:[[self class] instanceMethodSignatureForSelector:sel]];
